+++ Auth/auth.d.ts
@@ -1,0 +1,11 @@
+declare module "nativescript-azure-mobile-apps/user" {
+
+    export class MobileServiceUser {
+        constructor(o: { userId: string; });
+
+        setUserId(userId: string): void;
+        setAuthenticationToken(authenticationToken: string): void;
+        getUserId(): string;
+        getAuthenticationToken(): string;
+    }
+}
